+++ internal/commands/commands.go
@@ -17,15 +17,16 @@ type CommandDetails struct {
 }
 
 var commandRegistry map[string]CommandDetails = map[string]CommandDetails{
-	"CreateTables":   {"Create tables required for tracking activity over time", func() Command { return &CreateTables{} }},
-	"Exec":           {"Execute SQL statement across all DBs provided", func() Command { return &Exec{} }},
-	"Help":           {"Output usage", func() Command { return &Help{} }},
-	"NewActivity":    {"Output New Queries in the specified duration", func() Command { return &NewActivity{} }},
-	"QueryRow":       {"Query (single row) to execute across all DBs provided", func() Command { return &QueryRow{} }},
-	"QueryRows":      {"Query (multiple rows) to execute across all DBs provided", func() Command { return &QueryRows{} }},
-	"ResetIndexData": {"Reset index data", func() Command { return &ResetIndexData{} }},
-	"Snapshot":       {"Snapshot statistics tables", func() Command { return &Snapshot{} }},
-	"Summary":        {"Status summary", func() Command { return &Summary{} }},
+	"Exec":              {"Execute SQL statement across all DBs provided", func() Command { return &Exec{} }},
+	"Help":              {"Output usage", func() Command { return &Help{} }},
+	"NewActivity":       {"Output New Queries in the specified duration", func() Command { return &NewActivity{} }},
+	"QueryRow":          {"Query (single row) to execute across all DBs provided", func() Command { return &QueryRow{} }},
+	"QueryRows":         {"Query (multiple rows) to execute across all DBs provided", func() Command { return &QueryRows{} }},
+	"MonitorInitialize": {"Initialize infrastructure for activity monitoring", func() Command { return &MonitorInitialize{} }},
+	"MonitorReset":      {"Reset activity monitoring data", func() Command { return &MonitorReset{} }},
+	"MonitorTerminate":  {"Delete infrastructure for activity monitoring", func() Command { return &MonitorTerminate{} }},
+	"Snapshot":          {"Snapshot statistics tables", func() Command { return &Snapshot{} }},
+	"Summary":           {"Status summary", func() Command { return &Summary{} }},
 }
 
 var StatsTables = [...]string{"pg_stat_user_indexes", "pg_statio_user_indexes", "pg_stat_user_tables", "pg_statio_user_tables", "pg_stat_statements"}